# src/main/java/com/projet4/model/Utilisateur.java
package com.projet4.model;

import com.projet4.model.enums.Role;
import jakarta.persistence.*;

@Entity
@Table(name = "utilisateurs")
public class Utilisateur {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String name;
    private String email;
    private String password;

    @Enumerated(EnumType.STRING)
    private Role role;

    /**
     * Constructeur sans argument requis par JPA.
     * La spécification JPA impose la présence d’un constructeur no-arg public ou protected
     * pour l’instanciation via réflexion. :contentReference[oaicite:0]{index=0}
     */
    public Utilisateur() {
    }

    /**
     * Constructeur avec tous les champs, pratique pour instancier rapidement.
     */
    public Utilisateur(Long id, String name, String email, String password, Role role) {
        this.id = id;
        this.name = name;
        this.email = email;
        this.password = password;
        this.role = role;
    }

    // Getters et setters JavaBeans

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public Role getRole() {
        return role;
    }

    public void setRole(Role role) {
        this.role = role;
    }
}